Move crate tests into `workspace_tests` crate.
