+++ Cargo.toml
@@ -25,4 +25,4 @@ tokio = { version = "1.11.0", features = ["fs", "rt", "rt-multi-thread", "sync",
 tokio-util = { version = "0.6.7", features = ["codec"] }
 
 [workspace]
-members = ["crate/*"]
+members = ["crate/*", "workspace_tests"]
