+++ crate/cfg_model/src/station_id_invalid_fmt.rs
@@ -32,20 +32,3 @@ impl<'s> fmt::Display for StationIdInvalidFmt<'s> {
 }
 
 impl<'s> std::error::Error for StationIdInvalidFmt<'s> {}
-
-#[cfg(test)]
-mod tests {
-    use std::borrow::Cow;
-
-    use super::StationIdInvalidFmt;
-
-    #[test]
-    fn display_returns_readable_message() {
-        let station_id_invalid_fmt = StationIdInvalidFmt::new(Cow::Borrowed("a b c"));
-
-        assert_eq!(
-            "`a b c` is not a valid station ID. Station IDs can only contain letters, numbers, and underscores.",
-            station_id_invalid_fmt.to_string()
-        );
-    }
-}
