+++ crate/cfg_model/src/station_spec.rs
@@ -100,24 +100,3 @@ impl<E> fmt::Display for StationSpec<E> {
         write!(f, "{}: {}", self.name, self.description)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::StationSpec;
-    use crate::{StationFn, StationId, StationIdInvalidFmt, StationSpecFns};
-
-    #[test]
-    fn display_returns_readable_informative_message() -> Result<(), StationIdInvalidFmt<'static>> {
-        let station_id = StationId::new("station_id")?;
-        let name = String::from("Station Name");
-        let description = String::from("One liner.");
-        let station_spec_fns = {
-            let visit_fn = StationFn::new(|_, _| Box::pin(async { Result::<(), ()>::Ok(()) }));
-            StationSpecFns::new(visit_fn)
-        };
-        let station_spec = StationSpec::new(station_id, name, description, station_spec_fns);
-
-        assert_eq!("Station Name: One liner.", station_spec.to_string());
-        Ok(())
-    }
-}
