+++ crate/rt_logic/Cargo.toml
@@ -3,12 +3,16 @@ name = "choochoo_rt_logic"
 version = "0.1.0"
 authors = ["Azriel Hoh <azriel91@gmail.com>"]
 edition = "2018"
-description = "Runtime data types for the choochoo automation library."
+description = "Runtime visit logic for the choochoo automation library."
 repository = "https://github.com/azriel91/choochoo"
-documentation = "https://docs.rs/choochoo_rt_model/"
+documentation = "https://docs.rs/choochoo_rt_logic/"
 readme = "../../README.md"
 keywords = ["automation"]
 license = "MIT OR Apache-2.0"
+
+[lib]
+doctest = false
+test = false
 
 [dependencies]
 choochoo_cfg_model = { path = "../cfg_model", version = "0.1.0" }
