+++ crate/rt_logic/src/lib.rs
@@ -1,4 +1,4 @@
-//! Logic to update runtime data.
+//! Runtime visit logic for the choochoo automation library.
 
 pub use crate::{driver::Driver, visit_status_updater::VisitStatusUpdater};
