+++ crate/rt_model/Cargo.toml
@@ -10,6 +10,10 @@ readme = "../../README.md"
 keywords = ["automation"]
 license = "MIT OR Apache-2.0"
 
+[lib]
+doctest = false
+test = false
+
 [dependencies]
 choochoo_cfg_model = { path = "../cfg_model", version = "0.1.0" }
 daggy = { version = "0.7.0", default-features = false }
