+++ workspace_tests/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "workspace_tests"
+version = "0.1.0"
+authors = ["Azriel Hoh <azriel91@gmail.com>"]
+edition = "2018"
+description = "Tests for the choochoo automation library."
+repository = "https://github.com/azriel91/choochoo"
+documentation = "https://docs.rs/choochoo/"
+readme = "../README.md"
+keywords = ["automation"]
+license = "MIT OR Apache-2.0"
+publish = false
+
+[lib]
+doctest = false
+test = true
+
+[dependencies]
+choochoo_cfg_model = { path = "../crate/cfg_model", version = "0.1.0" }
+choochoo_rt_logic = { path = "../crate/rt_logic", version = "0.1.0" }
+choochoo_rt_model = { path = "../crate/rt_model", version = "0.1.0" }
+tokio = { version = "1.11.0", features = ["io-util", "rt", "sync"] }
