+++ workspace_tests/src/lib.rs
@@ -1,0 +1,7 @@
+#![cfg(test)]
+
+//! Tests for the choochoo automation library.
+
+mod cfg_model;
+mod rt_logic;
+mod rt_model;
